+++ src/portaudio/ringbuffer.c
@@ -43,7 +43,7 @@
  * Initialize FIFO.
  * numBytes must be power of 2, returns -1 if not.
  */
-long RingBuffer_Init( RingBuffer *rbuf, long numBytes, void *dataPtr )
+int32_t RingBuffer_Init( RingBuffer *rbuf, int32_t numBytes, void *dataPtr )
 {
     if( ((numBytes-1) & numBytes) != 0) return -1; /* Not Power of two. */
     rbuf->bufferSize = numBytes;
@@ -55,13 +55,13 @@ long RingBuffer_Init( RingBuffer *rbuf, long numBytes, void *dataPtr )
 }
 /***************************************************************************
 ** Return number of bytes available for reading. */
-long RingBuffer_GetReadAvailable( RingBuffer *rbuf )
+int32_t RingBuffer_GetReadAvailable( RingBuffer *rbuf )
 {
     return ( (rbuf->writeIndex - rbuf->readIndex) & rbuf->bigMask );
 }
 /***************************************************************************
 ** Return number of bytes available for writing. */
-long RingBuffer_GetWriteAvailable( RingBuffer *rbuf )
+int32_t RingBuffer_GetWriteAvailable( RingBuffer *rbuf )
 {
     return ( rbuf->bufferSize - RingBuffer_GetReadAvailable(rbuf));
 }
@@ -79,19 +79,19 @@ void RingBuffer_Flush( RingBuffer *rbuf )
 ** If non-contiguous, size2 will be the size of second region.
 ** Returns room available to be written or numBytes, whichever is smaller.
 */
-long RingBuffer_GetWriteRegions( RingBuffer *rbuf, long numBytes,
-                                 void **dataPtr1, long *sizePtr1,
-                                 void **dataPtr2, long *sizePtr2 )
+int32_t RingBuffer_GetWriteRegions( RingBuffer *rbuf, int32_t numBytes,
+                                 void **dataPtr1, int32_t *sizePtr1,
+                                 void **dataPtr2, int32_t *sizePtr2 )
 {
-    long   index;
-    long   available = RingBuffer_GetWriteAvailable( rbuf );
+    int32_t   index;
+    int32_t   available = RingBuffer_GetWriteAvailable( rbuf );
     if( numBytes > available ) numBytes = available;
     /* Check to see if write is not contiguous. */
     index = rbuf->writeIndex & rbuf->smallMask;
     if( (index + numBytes) > rbuf->bufferSize )
     {
         /* Write data in two blocks that wrap the buffer. */
-        long   firstHalf = rbuf->bufferSize - index;
+        int32_t   firstHalf = rbuf->bufferSize - index;
         *dataPtr1 = &rbuf->buffer[index];
         *sizePtr1 = firstHalf;
         *dataPtr2 = &rbuf->buffer[0];
@@ -110,7 +110,7 @@ long RingBuffer_GetWriteRegions( RingBuffer *rbuf, long numBytes,
 
 /***************************************************************************
 */
-long RingBuffer_AdvanceWriteIndex( RingBuffer *rbuf, long numBytes )
+int32_t RingBuffer_AdvanceWriteIndex( RingBuffer *rbuf, int32_t numBytes )
 {
     return rbuf->writeIndex = (rbuf->writeIndex + numBytes) & rbuf->bigMask;
 }
@@ -121,19 +121,19 @@ long RingBuffer_AdvanceWriteIndex( RingBuffer *rbuf, long numBytes )
 ** If non-contiguous, size2 will be the size of second region.
 ** Returns room available to be written or numBytes, whichever is smaller.
 */
-long RingBuffer_GetReadRegions( RingBuffer *rbuf, long numBytes,
-                                void **dataPtr1, long *sizePtr1,
-                                void **dataPtr2, long *sizePtr2 )
+int32_t RingBuffer_GetReadRegions( RingBuffer *rbuf, int32_t numBytes,
+                                void **dataPtr1, int32_t *sizePtr1,
+                                void **dataPtr2, int32_t *sizePtr2 )
 {
-    long   index;
-    long   available = RingBuffer_GetReadAvailable( rbuf );
+    int32_t   index;
+    int32_t   available = RingBuffer_GetReadAvailable( rbuf );
     if( numBytes > available ) numBytes = available;
     /* Check to see if read is not contiguous. */
     index = rbuf->readIndex & rbuf->smallMask;
     if( (index + numBytes) > rbuf->bufferSize )
     {
         /* Write data in two blocks that wrap the buffer. */
-        long firstHalf = rbuf->bufferSize - index;
+        int32_t firstHalf = rbuf->bufferSize - index;
         *dataPtr1 = &rbuf->buffer[index];
         *sizePtr1 = firstHalf;
         *dataPtr2 = &rbuf->buffer[0];
@@ -150,16 +150,16 @@ long RingBuffer_GetReadRegions( RingBuffer *rbuf, long numBytes,
 }
 /***************************************************************************
 */
-long RingBuffer_AdvanceReadIndex( RingBuffer *rbuf, long numBytes )
+int32_t RingBuffer_AdvanceReadIndex( RingBuffer *rbuf, int32_t numBytes )
 {
     return rbuf->readIndex = (rbuf->readIndex + numBytes) & rbuf->bigMask;
 }
 
 /***************************************************************************
 ** Return bytes written. */
-long RingBuffer_Write( RingBuffer *rbuf, void *data, long numBytes )
+int32_t RingBuffer_Write( RingBuffer *rbuf, void *data, int32_t numBytes )
 {
-    long size1, size2, numWritten;
+    int32_t size1, size2, numWritten;
     void *data1, *data2;
     numWritten = RingBuffer_GetWriteRegions( rbuf, numBytes, &data1, &size1, &data2, &size2 );
     if( size2 > 0 )
@@ -179,9 +179,9 @@ long RingBuffer_Write( RingBuffer *rbuf, void *data, long numBytes )
 
 /***************************************************************************
 ** Return bytes read. */
-long RingBuffer_Read( RingBuffer *rbuf, void *data, long numBytes )
+int32_t RingBuffer_Read( RingBuffer *rbuf, void *data, int32_t numBytes )
 {
-    long size1, size2, numRead;
+    int32_t size1, size2, numRead;
     void *data1, *data2;
     numRead = RingBuffer_GetReadRegions( rbuf, numBytes, &data1, &size1, &data2, &size2 );
     if( size2 > 0 )
